[plugin.video.newson] 1.0.3
