+++ plugin.video.newson/resources/images/icon.png
